Refactoring EmbeddedScriptControl add logic into ScriptControl to fix xml loading without dependency mappings.
