+++ SOC/Core/Forms/Pages/EmbeddedScriptSetControl.cs
@@ -75,7 +75,7 @@ namespace SOC.UI
 
             try
             {
-                Insert(ScriptDetails.LoadFromXml(loadFile.FileName));
+                ParentControl.Add(ScriptDetails.LoadFromXml(loadFile.FileName));
             }
             catch (Exception ex)
             {
@@ -169,156 +169,6 @@ namespace SOC.UI
         private void UpdateExportEnabled(bool newCheckboxValue)
         {
             buttonExportVariablesScripts.Enabled = checkedListBoxVariables.CheckedItems.Count + checkedListBoxScripts.CheckedItems.Count + (newCheckboxValue ? 1 : -1) > 0;
-        }
-
-        private void Insert(ScriptDetails incomingScriptDetails)
-        {
-            var variables = incomingScriptDetails.VariableDeclarations;
-
-            var scripts = incomingScriptDetails.QStep_Main;
-
-            var scriptals = scripts
-                .SelectMany(script => script.Preconditionals.Concat(script.Operationals))
-                .ToList();
-            MapChoicesToTokens(scriptals);
-
-            var choices = scriptals.SelectMany(scriptal => scriptal.Choices).ToList();
-
-            VariableNode[] incomingNodes = ConvertToVariableNodesWithDependencies(variables, choices);
-
-            MapUnresolvedDependenciesFromControl(choices);
-            
-            VariableNode[] bestGuessNodes = CreateVariableNodesForUnresolvedDependencies(choices);
-
-            MapUnresolvedDependenciesToNil(choices);
-
-            AddToControl(incomingNodes);
-
-            if (bestGuessNodes.Length > 0)
-            {
-                MessageBox.Show($"Notice: The imported script(s) depended on {bestGuessNodes.Length} variable(s) that did not correspond to any imported or pre-existing variable(s).\n\nThe new dependency variable(s) have been extrapolated.", "Extrapolated Variable(s) Created", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                AddToControl(bestGuessNodes);
-            }
-
-            AddToControl(scripts);
-            ParentControl.RedrawScriptDependents();
-        }
-
-        private void MapChoicesToTokens(List<Scriptal> scriptals)
-        {
-            foreach (var scriptal in scriptals)
-                scriptal.TryMapChoicesToTokens(out _);
-        }
-
-        private VariableNode[] ConvertToVariableNodesWithDependencies(List<LuaTableEntry> variables, List<Choice> choices)
-        {
-            VariableNode[] variableNodes = new VariableNode[variables.Count];
-
-            foreach (var variable in variables)
-            {
-                string name = variable.Key.Value.Trim('"');
-                if (ParentControl.VariableNameExists(name, ParentControl.treeViewVariables.Nodes))
-                {
-                    name = ParentControl.GetUniqueVariableName(name);
-                }
-
-                var variableNode = new VariableNode(Lua.TableEntry(name, variable.Value));
-
-                foreach (var choice in choices)
-                {
-                    if (choice.DependencyNameMatches(variable) && 
-                        choice.CorrespondingRuntimeToken.Allows(variable.Value, out _))
-                    {
-                        choice.SetVarNodeDependency(variableNode);
-                    }
-                }
-
-                variableNodes[variables.IndexOf(variable)] = variableNode;
-            }
-
-            return variableNodes;
-        }
-
-        private void MapUnresolvedDependenciesFromControl(List<Choice> choices)
-        {
-            foreach (VariableNode variableNode in ParentControl.treeViewVariables.Nodes)
-            {
-                foreach (var choice in choices)
-                {
-                    if (choice.Dependency == null && 
-                        choice.DependencyNameMatches(variableNode.GetEntry()) &&
-                        choice.CorrespondingRuntimeToken.Allows(variableNode.Entry.Value, out _))
-                    {
-                        choice.SetVarNodeDependency(variableNode);
-                    }
-                }
-            }
-        }
-
-        private VariableNode[] CreateVariableNodesForUnresolvedDependencies(List<Choice> choices)
-        {
-            List<VariableNode> extrapolatedNodes = new List<VariableNode>();
-
-            foreach (var choice in choices)
-            {
-                if (choice.Key == ScriptControl.CUSTOM_VARIABLE_SET && choice.Dependency == null && choice.Value is LuaTableIdentifier choiceIdentifier)
-                {
-                    bool matchesExistingExtrapolatedNode = false;
-                    foreach (VariableNode extrapolatedNode in extrapolatedNodes)
-                    {
-                        if (choice.DependencyNameMatches(extrapolatedNode.GetEntry()))
-                        {
-                            if (choiceIdentifier.EvaluatesTo == extrapolatedNode.GetEntry().Value.Type)
-                            {
-                                choice.SetVarNodeDependency(extrapolatedNode);
-                                matchesExistingExtrapolatedNode = true;
-                                break;
-                            } 
-                            else
-                            {
-                                choiceIdentifier.IdentifierKeys[0] = Lua.String(choiceIdentifier.IdentifierKeys[0].Value.Trim('"') + $"_{choiceIdentifier.EvaluatesTo}");
-                                break;
-                            }
-                        }
-                    }
-
-                    if (matchesExistingExtrapolatedNode) 
-                        continue;
-
-                    VariableNode ExtrapolatedNode = VariableNode.ExtrapolateDefault(choiceIdentifier);
-                    if (ExtrapolatedNode != null)
-                    {
-                        choice.SetVarNodeDependency(ExtrapolatedNode);
-                        extrapolatedNodes.Add(ExtrapolatedNode);
-                    }
-                }
-            }
-
-            return extrapolatedNodes.ToArray();
-        }
-
-        private void MapUnresolvedDependenciesToNil(List<Choice> choices)
-        {
-            foreach (var choice in choices)
-            {
-                if (choice.Key == ScriptControl.CUSTOM_VARIABLE_SET && choice.Dependency == null)
-                {
-                    choice.Value = new LuaNil();
-                }
-            }
-        }
-
-        private void AddToControl(VariableNode[] incomingNodes)
-        {
-            ParentControl.treeViewVariables.Nodes.AddRange(incomingNodes);
-        }
-
-        private void AddToControl(List<Script> scripts)
-        {
-            foreach (var script in scripts)
-            {
-                ParentControl.ScriptTablesRootNode.QStep_Main.Add(script);
-            }
         }
 
         private void checkedListBoxVariables_SelectedIndexChanged(object sender, EventArgs e)
